Organized the -Blocks classes into a .blocks package
