+++ src/main/java/com/brownian/thedeeps/blocks/DummyFrameBlock.java
@@ -1,9 +1,9 @@
-package com.brownian.thedeeps;
+package com.brownian.thedeeps.blocks;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 
-class DummyFrameBlock extends Block {
+public class DummyFrameBlock extends Block {
     public DummyFrameBlock() {
         super(Properties.copy(Blocks.OBSIDIAN));
     }
